+++ primitives/src/merkle_tree/prelude.rs
@@ -10,9 +10,9 @@ pub use crate::{
     impl_to_traversal_path_biguint, impl_to_traversal_path_primitives,
     merkle_tree::{
         append_only::MerkleTree, universal_merkle_tree::UniversalMerkleTree,
-        AppendableMerkleTreeScheme, DigestAlgorithm, Element, ForgetableMerkleTreeScheme, Index,
-        LookupResult, MerkleCommitment, MerkleTreeScheme, NodeValue, ToTraversalPath,
-        UniversalMerkleTreeScheme,
+        AppendableMerkleTreeScheme, DigestAlgorithm, Element, ForgetableMerkleTreeScheme,
+        ForgetableUniversalMerkleTreeScheme, Index, LookupResult, MerkleCommitment,
+        MerkleTreeScheme, NodeValue, ToTraversalPath, UniversalMerkleTreeScheme,
     },
 };
 
@@ -20,6 +20,8 @@ use crate::rescue::{sponge::RescueCRHF, RescueParameter};
 use ark_std::marker::PhantomData;
 use num_bigint::BigUint;
 use typenum::U3;
+
+use super::light_weight::LightWeightMerkleTree;
 
 /// Wrapper for rescue hash function
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -40,6 +42,9 @@ impl<F: RescueParameter> DigestAlgorithm<F, u64, F> for RescueHash<F> {
 
 /// A standard merkle tree using RATE-3 rescue hash function
 pub type RescueMerkleTree<F> = MerkleTree<F, RescueHash<F>, u64, U3, F>;
+
+/// A standard light merkle tree using RATE-3 rescue hash function
+pub type RescueLightWeightMerkleTree<F> = LightWeightMerkleTree<F, RescueHash<F>, u64, U3, F>;
 
 impl<F: RescueParameter> DigestAlgorithm<F, BigUint, F> for RescueHash<F> {
     fn digest(data: &[F]) -> F {